watson_stt_app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ibm-watson
ibm-cos-sdk
ffmpeg-python

